requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
openai==1.30.0
python-dotenv==1.0.1
numpy==1.26.4